file-watcher cmd line added. using rubys sleep

+++ project2/lib/file-watch/file-watch-contract.rb
@@ -6,7 +6,7 @@ module ContractFileWatch
   
   def pre_FileWatch(mode, duration, *files)
     assert(mode.respond_to?(:to_s) && !mode.strip.empty?)
-    assert(duration.respond_to? :to_i)
+    assert(duration.respond_to? :to_f)
     assert !files.empty?, "File names must be given."
     files.each do |f|
       assert(f.respond_to?(:to_s) && !f.strip.empty?, "File strings cannot be empty")
